Move falsify_or_delete_celo_pending_epoch_operation to respective chain module
